Fixed all the warnings, enabled warnings as errors

+++ libvgmtool/gd3.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstdint>
 
 // GD3 tag format definitions
 // and functionality
@@ -9,8 +10,8 @@ class IVGMToolCallback;
 struct TGD3Header
 {
     char id_string[4]; // "Gd3 "
-    long version; // 0x000000100 for 1.00
-    long length; // Length of string data following this point
+    uint32_t version; // 0x000000100 for 1.00
+    uint32_t length; // Length of string data following this point
 };
 
 #define NumGD3Strings 11
